+++ tests/captured_data/MCU680.hexdump
