+++ tests/captured_data/PMS3003.hexdump
