+++ tests/captured_data/PMS5003T.hexdump
