+++ tests/captured_data/PMSx003.hexdump
