+++ tests/captured_data/SDS01x.hexdump
