+++ tests/sensor/test_message.py
@@ -1,13 +1,11 @@
 import pytest
 
-from pms import SensorWarning, logger
+from pms import SensorWarning
 from pypms.bosch_sensortec import mcu680
 from pypms.honeywell import hpma115s0
 from pypms.novafitness import sds01x
 from pypms.plantower import pms3003, pmsx003
 from pypms.senserion import sps30
-
-logger.setLevel("DEBUG")
 
 
 @pytest.mark.parametrize(
